prevent auto increment serial counter
we use try/except around insert to determine if the
record exists. if it does, we get an IntegrityError,
and then we go fetch the record. however, when we do
this, postgresql will increment the serial counter,
even though no record is inserted.

we have found that:
1. the vast majority of the time, we are fetching an
existing record. because of this, doing the lookup
first yields faster results since we'er doing 1
operation instead of 2. in addition, it's better if
the majority of our operations do not intentially
trigger postgresql exception, which is more costly
than not.

2. we did not like the asthetics of the serial id
field jumping by large numbers. it was harder to
quickly glance and get estimates for the number of
records we have at any given point.